+++ R/class-workbook-wrappers.R
@@ -2007,3 +2007,60 @@ wb_add_border <- function(
   )
 
 }
+
+#' add fill for fill region
+#'
+#' @description wb wrapper to create fill for cell region
+#' @param wb a workbook
+#' @param sheet the worksheet
+#' @param dims the cell range
+#' @param color the colors to apply, e.g. yellow: c(rgb = "FFFFFF00")
+#' @param pattern various default "none" but others are possible:
+#'  "solid", "mediumGray", "darkGray", "lightGray", "darkHorizontal",
+#'  "darkVertical", "darkDown", "darkUp", "darkGrid", "darkTrellis",
+#'  "lightHorizontal", "lightVertical", "lightDown", "lightUp", "lightGrid",
+#'  "lightTrellis", "gray125", "gray0625"
+#' @param gradient_fill a gradient fill xml pattern.
+#' @param every_nth_col which col should be filled
+#' @param every_nth_row which row should be filled
+#' @examples
+#' wb <- wb_workbook() %>% wb_add_worksheet("S1") %>% wb_add_data("S1", mtcars)
+#' wb <- wb %>% wb_add_fill("S1", dims = "D5:J23", color = c(rgb = "FFFFFF00"))
+#' wb <- wb %>% wb_add_fill("S1", dims = "B22:D27", color = c(rgb = "FF00FF00"))
+#'
+#' wb <- wb %>%  wb_add_worksheet("S2") %>% wb_add_data("S2", mtcars)
+#'
+#' gradient_fill1 <- '<gradientFill degree="90">
+#' <stop position="0"><color rgb="FF92D050"/></stop>
+#' <stop position="1"><color rgb="FF0070C0"/></stop>
+#' </gradientFill>'
+#' wb <- wb %>% wb_add_fill("S2", dims = "A2:K5", gradient_fill = gradient_fill1)
+#'
+#' gradient_fill2 <- '<gradientFill type="path" left="0.2" right="0.8" top="0.2" bottom="0.8">
+#' <stop position="0"><color theme="0"/></stop>
+#' <stop position="1"><color theme="4"/></stop>
+#' </gradientFill>'
+#' wb <- wb %>% wb_add_fill("S2", dims = "A7:K10", gradient_fill = gradient_fill2)
+#' @return The `wbWorksheetObject`, invisibly
+#' @export
+wb_add_fill <- function(
+    wb,
+    sheet,
+    dims,
+    color = "",
+    pattern = "solid",
+    gradient_fill = "",
+    every_nth_col = 1,
+    every_nth_row = 1
+) {
+  sheet <- wb_validate_sheet(wb, sheet)
+  wb$clone()$add_fill(
+    sheet = sheet,
+    dims = dims,
+    color = color,
+    pattern = pattern,
+    gradient_fill = gradient_fill,
+    every_nth_col = every_nth_col,
+    every_nth_row = every_nth_row
+  )
+}