+++ R/class-workbook.R
@@ -4135,7 +4135,7 @@ wbWorkbook <- R6::R6Class(
     #' wb <- wb_workbook()
     #' wb$add_worksheet("S1")$add_data("S1", mtcars)
     #' wb$add_border(1, dims = "A2:K33", inner_vgrid = "thin", inner_vcolor = c(rgb="FF808080"))
-    #'
+    #' @return The `wbWorksheetObject`, invisibly
     add_border = function(
       sheet         = 1,
       dims          = "A1",
@@ -4381,8 +4381,8 @@ wbWorkbook <- R6::R6Class(
           self$styles_mgr$add(full_single, sfull_single)
           xf_prev <- get_cell_styles(self, sheet, dims)
           xf_full_single <- set_border(xf_prev, self$styles_mgr$get_border_id(sfull_single))
-          self$styles_mgr$add(xf_full_single, s(xf_full_single))
-          set_cell_style(self, sheet, dims, self$styles_mgr$get_xf_id(s(xf_full_single)))
+          self$styles_mgr$add(xf_full_single, xf_full_single)
+          set_cell_style(self, sheet, dims, self$styles_mgr$get_xf_id(xf_full_single))
         }
 
         # create top & bottom piece
@@ -4392,15 +4392,15 @@ wbWorkbook <- R6::R6Class(
           self$styles_mgr$add(top_single, stop_single)
           xf_prev <- get_cell_styles(self, sheet, dim_top_single)
           xf_top_single <- set_border(xf_prev, self$styles_mgr$get_border_id(stop_single))
-          self$styles_mgr$add(xf_top_single, s(xf_top_single))
-          set_cell_style(self, sheet, dim_top_single, self$styles_mgr$get_xf_id(s(xf_top_single)))
+          self$styles_mgr$add(xf_top_single, xf_top_single)
+          set_cell_style(self, sheet, dim_top_single, self$styles_mgr$get_xf_id(xf_top_single))
 
           # bottom single
           self$styles_mgr$add(bottom_single, sbottom_single)
           xf_prev <- get_cell_styles(self, sheet, dim_bottom_single)
           xf_bottom_single <- set_border(xf_prev, self$styles_mgr$get_border_id(sbottom_single))
-          self$styles_mgr$add(xf_bottom_single, s(xf_bottom_single))
-          set_cell_style(self, sheet, dim_bottom_single, self$styles_mgr$get_xf_id(s(xf_bottom_single)))
+          self$styles_mgr$add(xf_bottom_single, xf_bottom_single)
+          set_cell_style(self, sheet, dim_bottom_single, self$styles_mgr$get_xf_id(xf_bottom_single))
         }
 
         # create middle piece(s)
@@ -4410,8 +4410,8 @@ wbWorkbook <- R6::R6Class(
           self$styles_mgr$add(middle_single, smiddle_single)
           xf_prev <- get_cell_styles(self, sheet, dim_middle_single)
           xf_middle_single <- set_border(xf_prev, self$styles_mgr$get_border_id(smiddle_single))
-          self$styles_mgr$add(xf_middle_single, s(xf_middle_single))
-          set_cell_style(self, sheet, dim_middle_single, self$styles_mgr$get_xf_id(s(xf_middle_single)))
+          self$styles_mgr$add(xf_middle_single, xf_middle_single)
+          set_cell_style(self, sheet, dim_middle_single, self$styles_mgr$get_xf_id(xf_middle_single))
         }
 
       }
@@ -4423,15 +4423,15 @@ wbWorkbook <- R6::R6Class(
         self$styles_mgr$add(left_single, sleft_single)
         xf_prev <- get_cell_styles(self, sheet, dim_left_single)
         xf_left_single <- set_border(xf_prev, self$styles_mgr$get_border_id(sleft_single))
-        self$styles_mgr$add(xf_left_single, s(xf_left_single))
-        set_cell_style(self, sheet, dim_left_single, self$styles_mgr$get_xf_id(s(xf_left_single)))
+        self$styles_mgr$add(xf_left_single, xf_left_single)
+        set_cell_style(self, sheet, dim_left_single, self$styles_mgr$get_xf_id(xf_left_single))
 
         # right single
         self$styles_mgr$add(right_single, sright_single)
         xf_prev <- get_cell_styles(self, sheet, dim_right_single)
         xf_right_single <- set_border(xf_prev, self$styles_mgr$get_border_id(sright_single))
-        self$styles_mgr$add(xf_right_single, s(xf_right_single))
-        set_cell_style(self, sheet, dim_right_single, self$styles_mgr$get_xf_id(s(xf_right_single)))
+        self$styles_mgr$add(xf_right_single, xf_right_single)
+        set_cell_style(self, sheet, dim_right_single, self$styles_mgr$get_xf_id(xf_right_single))
 
         # add single center piece(s)
         if (ncol(df) >= 3) {
@@ -4440,8 +4440,8 @@ wbWorkbook <- R6::R6Class(
           self$styles_mgr$add(center_single, scenter_single)
           xf_prev <- get_cell_styles(self, sheet, dim_center_single)
           xf_center_single <- set_border(xf_prev, self$styles_mgr$get_border_id(scenter_single))
-          self$styles_mgr$add(xf_center_single, s(xf_center_single))
-          set_cell_style(self, sheet, dim_center_single, self$styles_mgr$get_xf_id(s(xf_center_single)))
+          self$styles_mgr$add(xf_center_single, xf_center_single)
+          set_cell_style(self, sheet, dim_center_single, self$styles_mgr$get_xf_id(xf_center_single))
         }
 
       }
@@ -4453,29 +4453,29 @@ wbWorkbook <- R6::R6Class(
         self$styles_mgr$add(top_left, stop_left)
         xf_prev <- get_cell_styles(self, sheet, dim_top_left)
         xf_top_left <- set_border(xf_prev, self$styles_mgr$get_border_id(stop_left))
-        self$styles_mgr$add(xf_top_left, s(xf_top_left))
-        set_cell_style(self, sheet, dim_top_left, self$styles_mgr$get_xf_id(s(xf_top_left)))
+        self$styles_mgr$add(xf_top_left, xf_top_left)
+        set_cell_style(self, sheet, dim_top_left, self$styles_mgr$get_xf_id(xf_top_left))
 
         # top right
         self$styles_mgr$add(top_right, stop_right)
         xf_prev <- get_cell_styles(self, sheet, dim_top_right)
         xf_top_right <- set_border(xf_prev, self$styles_mgr$get_border_id(stop_right))
-        self$styles_mgr$add(xf_top_right, s(xf_top_right))
-        set_cell_style(self, sheet, dim_top_right, self$styles_mgr$get_xf_id(s(xf_top_right)))
+        self$styles_mgr$add(xf_top_right, xf_top_right)
+        set_cell_style(self, sheet, dim_top_right, self$styles_mgr$get_xf_id(xf_top_right))
 
         # bottom left
         self$styles_mgr$add(bottom_left, sbottom_left)
         xf_prev <- get_cell_styles(self, sheet, dim_bottom_left)
         xf_bottom_left <- set_border(xf_prev, self$styles_mgr$get_border_id(sbottom_left))
-        self$styles_mgr$add(xf_bottom_left, s(xf_bottom_left))
-        set_cell_style(self, sheet, dim_bottom_left, self$styles_mgr$get_xf_id(s(xf_bottom_left)))
+        self$styles_mgr$add(xf_bottom_left, xf_bottom_left)
+        set_cell_style(self, sheet, dim_bottom_left, self$styles_mgr$get_xf_id(xf_bottom_left))
 
         # bottom right
         self$styles_mgr$add(bottom_right, sbottom_right)
         xf_prev <- get_cell_styles(self, sheet, dim_bottom_right)
         xf_bottom_right <- set_border(xf_prev, self$styles_mgr$get_border_id(sbottom_right))
-        self$styles_mgr$add(xf_bottom_right, s(xf_bottom_right))
-        set_cell_style(self, sheet, dim_bottom_right, self$styles_mgr$get_xf_id(s(xf_bottom_right)))
+        self$styles_mgr$add(xf_bottom_right, xf_bottom_right)
+        set_cell_style(self, sheet, dim_bottom_right, self$styles_mgr$get_xf_id(xf_bottom_right))
       }
 
       # create left and right middle pieces
@@ -4485,15 +4485,15 @@ wbWorkbook <- R6::R6Class(
         self$styles_mgr$add(middle_left, smiddle_left)
         xf_prev <- get_cell_styles(self, sheet, dim_middle_left)
         xf_middle_left <- set_border(xf_prev, self$styles_mgr$get_border_id(smiddle_left))
-        self$styles_mgr$add(xf_middle_left, s(xf_middle_left))
-        set_cell_style(self, sheet, dim_middle_left, self$styles_mgr$get_xf_id(s(xf_middle_left)))
+        self$styles_mgr$add(xf_middle_left, xf_middle_left)
+        set_cell_style(self, sheet, dim_middle_left, self$styles_mgr$get_xf_id(xf_middle_left))
 
         # middle right
         self$styles_mgr$add(middle_right, smiddle_right)
         xf_prev <- get_cell_styles(self, sheet, dim_middle_right)
         xf_middle_right <- set_border(xf_prev, self$styles_mgr$get_border_id(smiddle_right))
-        self$styles_mgr$add(xf_middle_right, s(xf_middle_right))
-        set_cell_style(self, sheet, dim_middle_right, self$styles_mgr$get_xf_id(s(xf_middle_right)))
+        self$styles_mgr$add(xf_middle_right, xf_middle_right)
+        set_cell_style(self, sheet, dim_middle_right, self$styles_mgr$get_xf_id(xf_middle_right))
       }
 
       # create top and bottom center pieces
@@ -4503,15 +4503,15 @@ wbWorkbook <- R6::R6Class(
         self$styles_mgr$add(top_center, stop_center)
         xf_prev <- get_cell_styles(self, sheet, dim_top_center)
         xf_top_center <- set_border(xf_prev, self$styles_mgr$get_border_id(stop_center))
-        self$styles_mgr$add(xf_top_center, s(xf_top_center))
-        set_cell_style(self, sheet, dim_top_center, self$styles_mgr$get_xf_id(s(xf_top_center)))
+        self$styles_mgr$add(xf_top_center, xf_top_center)
+        set_cell_style(self, sheet, dim_top_center, self$styles_mgr$get_xf_id(xf_top_center))
 
         # bottom center
         self$styles_mgr$add(bottom_center, sbottom_center)
         xf_prev <- get_cell_styles(self, sheet, dim_bottom_center)
         xf_bottom_center <- set_border(xf_prev, self$styles_mgr$get_border_id(sbottom_center))
-        self$styles_mgr$add(xf_bottom_center, s(xf_bottom_center))
-        set_cell_style(self, sheet, dim_bottom_center, self$styles_mgr$get_xf_id(s(xf_bottom_center)))
+        self$styles_mgr$add(xf_bottom_center, xf_bottom_center)
+        set_cell_style(self, sheet, dim_bottom_center, self$styles_mgr$get_xf_id(xf_bottom_center))
       }
 
       if (nrow(df) > 2 && ncol(df) > 2) {
@@ -4520,8 +4520,70 @@ wbWorkbook <- R6::R6Class(
         self$styles_mgr$add(inner_cell, sinner_cell)
         xf_prev <- get_cell_styles(self, sheet, dim_inner_cell)
         xf_inner_cell <- set_border(xf_prev, self$styles_mgr$get_border_id(sinner_cell))
-        self$styles_mgr$add(xf_inner_cell, s(xf_inner_cell))
-        set_cell_style(self, sheet, dim_inner_cell, self$styles_mgr$get_xf_id(s(xf_inner_cell)))
+        self$styles_mgr$add(xf_inner_cell, xf_inner_cell)
+        set_cell_style(self, sheet, dim_inner_cell, self$styles_mgr$get_xf_id(xf_inner_cell))
+      }
+
+      return(self)
+    },
+
+    #' @description provide simple fill function
+    #' @param sheet the worksheet
+    #' @param dims the cell range
+    #' @param color the colors to apply, e.g. yellow: c(rgb = "FFFFFF00")
+    #' @param pattern various default "none" but others are possible:
+    #'  "solid", "mediumGray", "darkGray", "lightGray", "darkHorizontal",
+    #'  "darkVertical", "darkDown", "darkUp", "darkGrid", "darkTrellis",
+    #'  "lightHorizontal", "lightVertical", "lightDown", "lightUp", "lightGrid",
+    #'  "lightTrellis", "gray125", "gray0625"
+    #' @param gradient_fill a gradient fill xml pattern.
+    #' @param every_nth_col which col should be filled
+    #' @param every_nth_row which row should be filled
+    #' @examples
+    #'  # example from the gradient fill manual page
+    #'  gradient_fill <- "<gradientFill degree=\"90\">
+    #'    <stop position=\"0\"><color rgb=\"FF92D050\"/></stop>
+    #'    <stop position=\"1\"><color rgb=\"FF0070C0\"/></stop>
+    #'   </gradientFill>"
+    #' @return The `wbWorksheetObject`, invisibly
+    add_fill = function(
+      sheet,
+      dims,
+      color = "",
+      pattern = "solid",
+      gradient_fill = "",
+      every_nth_col = 1,
+      every_nth_row = 1
+    ) {
+
+      new_fill <- create_fill(
+        gradientFill = gradient_fill,
+        patternType = pattern,
+        fgColor = color
+      )
+
+      smp <- paste0(sample(letters, size = 6, replace = TRUE), collapse = "")
+      snew_fill <- paste0(smp, "new_fill")
+      sxf_new_fill <- paste0(smp, "xf_new_fill")
+
+
+      self$styles_mgr$add(new_fill, snew_fill)
+
+      # dim in dataframe can contain various styles. go cell by cell.
+      did <- dims_to_dataframe(dims, fill = TRUE)
+      # select a few cols and rows to fill
+      cols <- (seq_len(ncol(did)) %% every_nth_col) == 0
+      rows <- (seq_len(nrow(did)) %% every_nth_row) == 0
+
+      dims <- unname(unlist(did[rows, cols, drop = FALSE]))
+
+      for (dim in dims) {
+        sxf_new_fill_x <- paste0(sxf_new_fill, which(dims %in% dim))
+        xf_prev <- get_cell_styles(self, sheet, dim)
+        xf_new_fill <- set_fill(xf_prev, self$styles_mgr$get_fill_id(snew_fill))
+        self$styles_mgr$add(xf_new_fill, xf_new_fill)
+        s_id <- self$styles_mgr$get_xf_id(xf_new_fill)
+        set_cell_style(self, sheet, dim, s_id)
       }
 
       return(self)
@@ -5354,7 +5416,7 @@ wb_get_sheet_name = function(wb, index = NULL) {
   if (any(index > n)) {
     stop("Invalid sheet index. Workbook ", n, " sheet(s)", call. = FALSE)
   }
-  
+
   # keep index 0 as ""
   z <- vector("character", length(index))
   names(z) <- index